Prettify path to page object definition
In steps.d.ts, because CodeceptJS generates the file from the root
directory (why?), the import path to the page object definition is
relative to the root. VS Code doesn't like this, although it builds and
runs fine. However, we can make the path relative to the correctly
placed steps.d.ts to appease VS Code without breaking anything.