Use "export ="
Instead of the somewhat clumsy use of "module.exports =" with a named
object and type, use "export =" and the type name defined by steps.d.ts.
That type name, being autogenerated, is annoyingly unconventional, but
it works and leaves significantly fewer moving parts in both the page
objects and the test files.