Monkey-patch REST helper members
This commit demonstrates working with the REST helper.

The REST helper's type definitions are completely off, making the helper
useless from TypeScript [1]. The members hare trivially incomplete JSDoc
specifications, which is the source of the error; but the types require
a third party library and the tooling does not have a way to pull in
those third party types. I don't know of a way to accomplish that
manually. Consequently, monkey-patching the helper's members locally
ends up being easier.

Besides fixing the upstream type definitions, this commit also
demonstrates various ways to structure TypeScript and tests for
flexibility and ergonomics: there is (relative) type safety of both
inputs and outputs as well as a dedicated API "page" object that hides
the low-level request details.

Note that there is no demonstration of error responses. This could
possibly be accomplished with a union type over the success and error
resopnse.

[1] codeceptjs/CodeceptJS#2590