Only load Puppeteer helper when necessary
The Puppeteer helper starts a browser on every test. This costs easily
500-700 ms. The REST helper doesn't use Puppeteer at all, to the point
that the URL is not even loaded in the browser. That process overhead is
pure busy wait, as well as confusing w.r.t. what is really happening
("why is the page blank?").

It is possible to load only strictly needed helpers; this demonstrates
how. Note, though, that when generating types all helpers must be
loaded, otherwise their types will not be included.

At the end of the day, this may well be a terrible idea...